pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spotty"
version = "0.1.5"
description = "Home Assistant NFC Bridge for PN532 readers"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "requests>=2.28.0",
    "gpiozero>=1.6.0",
    "spidev>=3.0.0",
    "pyserial>=3.5",
    "pyyaml>=6.0",
    "paho-mqtt>=2.0.0",
]

[project.scripts]
spotty = "spotty.main:main"
